apkh-search/requirements.txt
fastapi
uvicorn[standard]
PyJWT
cryptography
httpx
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain-google-genai
langchain-anthropic
google-genai
openai
anthropic
beautifulsoup4
PyMuPDF
pytesseract
Pillow
python-docx
openpyxl
tiktoken
langsmith
python-dotenv
python-multipart
